// frontend/src/types/langgraph.ts
export interface SearchItemsOptions {
    filter?: Record<string, any>;
    limit?: number;
    offset?: number;
    query?: string;
    refreshTtl?: boolean | null;
}

export interface StoreItem {
    namespace: string[];
    key: string;
    value: Record<string, any>;
    createdAt: string;
    updatedAt: string;
}

export interface SearchStoreItem extends StoreItem {
    score?: number;
}

export interface ListNamespaceOptions {
    prefix?: string[];
    suffix?: string[];
    maxDepth?: number;
    limit?: number;
    offset?: number;
}